requirements.txt
Django==6.0.3
python-dotenv==1.2.2
psycopg2==2.9.11
djangorestframework==3.16.1
django-filter==25.2
djangorestframework_simplejwt==5.5.1